% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_training.R
\name{create_training}
\alias{create_training}
\title{This package will create a function called create_training()}
\usage{
create_training(path, ...)
}
\description{
It's callback is at: inst/rstudio/templates/project/create_training.dcf
}
